L'equilibrista con la stella
The story is set in South Tyrol, in the north Italy, during the second World War. The lifes of a Jew girl, Anastasia, and a circus boy, Thomas, meet when the little circus of Freit family arrives in the city. They soon fall in love. But their love is full of sacrifices, because of the racial laws. So Anastasia decides to hide in the circus, but also there the situation is hard because of the son of the Circus Director, Artur, who hates her because of her Jewish origins and wants her out of the circus. At the same time all the family of Anastasia, staying at home is deported. Only the sister of Anastasia, Greta, will luckily survive. After a series of unfortunate events, Anastasia will be obliged to mask herself first, then to escape from the circus. The story continues trough thick and thin, following an impossible love story between Anastasia and Thomas.